Stage 4: verified uint8 mask cache and zero-I/O patch sampling (15x faster; data was 75-98% of step time)

+++ sailor/stage4/benchmark.py
@@ -91,8 +91,10 @@ def benchmark(model, sampler, batch_size: int = 4, steps: int = 60,
     for i in range(warmup + steps):
         t0 = time.perf_counter()
         x, y = sampler.batch(batch_size, epoch=i)
-        xb = torch.from_numpy(x).to(dev, non_blocking=True)
-        yb = torch.from_numpy(y).to(dev, non_blocking=True)
+        # Transfer as uint8, cast on device — 4x less PCIe traffic and the
+        # cast costs nothing on the GPU.
+        xb = torch.from_numpy(x).to(dev, non_blocking=True).float()
+        yb = torch.from_numpy(y).to(dev, non_blocking=True).float()
         if channels_last:
             xb = xb.to(memory_format=torch.channels_last_3d)
         if dev == "cuda":
